feat: Admin fee settings UI improvements and premium subscription enhancements
- Remove Add New Fee Setting button and all related creation functionality

- Remove delete actions from all fee settings tables

- Align all action buttons consistently across fee settings tables

- Add premium subscription expiration system with automated daily command

- Implement manual premium subscription renewal functionality

- Add default tokens for new users setting in admin panel

- Create trade history page with filtering and statistics

- Add premium/free badges to profile and navigation

- Improve token history page with transaction type differentiation

- Update email verification to work for unauthenticated users

- Enhance trade creation/edit with all user skills (registered + acquired)

- Prevent skill editing on trades with active requests

- Make premium users exempt from token fees for requests/acceptance

+++ app/Models/TradeFeeSetting.php
@@ -18,7 +18,7 @@ class TradeFeeSetting extends Model
 
     protected $casts = [
         'is_active' => 'boolean',
-        'fee_amount' => 'integer',
+        'fee_amount' => 'float', // Changed to float to support PHP prices with decimals
     ];
 
     /**
